move getApiUrl to api folder
